Make a directory to keep bidder page component. Edit bottomNav
