+++ client/src/components/ui/bidder/task-card.tsx
