+++ client/src/pages/component-showcase.tsx
@@ -1,6 +1,8 @@
 import { useState } from "react";
 
-import BottomNav from "@/components/ui/bottom-nav";
+import BottomNav from "@/components/ui/bidder/bottom-nav";
+import TaskCard from "@/components/ui/bidder/task-card";
+import TopNavtab from "@/components/ui/bidder/top-navtab";
 import { Button } from "@/components/ui/button";
 import {
   ErrorCallout,
@@ -30,8 +32,6 @@ import {
 } from "@/components/ui/inputs";
 import PersonDetail from "@/components/ui/person-detail";
 import ProfileTag from "@/components/ui/profile-tags";
-import TaskCard from "@/components/ui/task-card";
-import TopNavtab from "@/components/ui/top-navtab";
 
 function ComponentSection({
   title,
@@ -251,7 +251,7 @@ export default function ComponentShowcase() {
 
         <ComponentSection title="Bottom Nav">
           <div className="w-full">
-            <BottomNav isFixed={false} navIndex={0} />
+            <BottomNav isFixed={false} children={<></>} />
           </div>
         </ComponentSection>
